% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance_standard.R
\name{abundance_standart}
\alias{abundance_standart}
\title{abundance_standard}
\usage{
abundance_standart(viral_df, taxa = "Family", abuntype = "relative")
}
\arguments{
\item{viral_df}{dataframe, gene_to_genome.csv, output from VContact2}

\item{taxa}{taxon selected by the user among the following: "Genome","Order","Family","Genus". Default set to "Family"}

\item{abuntype}{abundance to construct the matrix. "relative" or "absolute"; default set to "relative"}
}
\description{
This is a function that suppresses log info
#'
}
\examples{
\dontrun{
quiet(cat("test"))
quiet(warning("test"))
quiet(warning("test"), all=T)
}
# This is a function that suppresses log info
}
