% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance_vc.R
\name{abundance_vc}
\alias{abundance_vc}
\title{abundance_vc}
\usage{
abundance_vc(
  viral_vc,
  taxa = "Family",
  output_type = "matrix",
  abuntype = "relative"
)
}
\arguments{
\item{viral_vc}{dataframe, gene_to_genome.csv, output from VContact2}

\item{taxa}{taxon selected by the user among the following: "Genome","Order","Family","Genus". Default set to "Family"}

\item{abuntype}{abundance to construct the matrix. "relative" or "absolute"; default set to "relative"}
}
\description{
Creates a abundance matrix from the genome_by_genome_overview.csv files outputed by VCONTACT2
}
\examples{
\dontrun{
quiet(cat("test"))
quiet(warning("test"))
quiet(warning("test"), all=T)
}
# This is a function that suppresses log info
}
