% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_statistics.R
\name{bio_statistics}
\alias{bio_statistics}
\title{bio_statistics}
\usage{
bio_statistics(abundance_mx)
}
\arguments{
\item{abundance_mx}{matrix with abundance values, output from vcontact2}
}
\description{
performs multiple diversity indexes
#'
}
\examples{
\dontrun{
quiet(cat("test"))
quiet(warning("test"))
quiet(warning("test"), all=T)
}
# This is a function that suppresses log info
}
