% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{cleaning}
\alias{cleaning}
\title{cleaning}
\usage{
cleaning(
  viral_data,
  output_from = "vcontact2",
  remove_flags = c("F", "T", "P", "A"),
  max_aux_score = 3,
  ...
)
}
\arguments{
\item{viral_data}{file to be clean}

\item{output_from}{vcontact2 or DRAMv}

\item{remove_flags}{flags to remove}

\item{max_aux_score}{maximun auxiliary score}

\item{...}{others}
}
\description{
cleans the output from vcontact2 or DRAM-v
#'
}
\examples{
\dontrun{
quiet(cat("test"))
quiet(warning("test"))
quiet(warning("test"), all=T)
}
# This is a function that suppresses log info
}
