% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conc_tsv.R
\name{conc_tsv}
\alias{conc_tsv}
\title{conc_tsv}
\usage{
conc_tsv(path)
}
\arguments{
\item{x}{the function to suppress the log info}

\item{all}{If TRUE then suppress warnings and messages as well; otherwise, only suppress printed output (such as from print or cat).}
}
\description{
This is a function that suppresses log info
#'
}
\examples{
\dontrun{
quiet(cat("test"))
quiet(warning("test"))
quiet(warning("test"), all=T)
}
# This is a function that suppresses log info
}
