% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_exploring.R
\name{db_exploring}
\alias{db_exploring}
\title{db_exploring}
\usage{
db_exploring(viral_annotations, database = "all")
}
\arguments{
\item{viral_annotations}{dataframe from DRAMv}

\item{database}{database to further analize}
}
\description{
isolates the database selected, counts for the biome and can perform other manipulations
#'
}
\examples{
\dontrun{
quiet(cat("test"))
quiet(warning("test"))
quiet(warning("test"), all=T)
}
# This is a function that suppresses log info
}
