% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/host_domain.R
\name{host_domain}
\alias{host_domain}
\title{host_domain}
\usage{
host_domain(vcontact_df, reduce_names = FALSE)
}
\arguments{
\item{vcontact_df}{dataframe from vcontact2}

\item{reduce_names}{if TRUE separates the Genome column so that only the host genus is present, default set to FALSE}
}
\description{
This function assigns to the host, or genome, the respective domain, archeae, bacteria or eukarya
#'
}
\examples{
\dontrun{
quiet(cat("test"))
quiet(warning("test"))
quiet(warning("test"), all=T)
}
# This is a function that suppresses log info
}
