% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_csv.R
\name{list_csv}
\alias{list_csv}
\title{list_csv}
\usage{
list_csv(flnm)
}
\arguments{
\item{flnm}{name of the file}
}
\description{
This is a function that suppresses log info
#'
}
\examples{
\dontrun{
quiet(cat("test"))
quiet(warning("test"))
quiet(warning("test"), all=T)
}
# This is a function that suppresses log info
}
