% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{cleaning}
\alias{cleaning}
\title{cleaning}
\usage{
cleaning(
  data,
  sepby = "-(?=[^-])",
  fullclean = TRUE,
  output_from = "vcontact2",
  ...
)
}
\arguments{
\item{sepby}{regex comands to separate the column. Vcontact2 regex: "-(?=[^-])" ; DRAM-v regex: "_(?=[^_]+$)"}
}
\description{
cleans
#'
}
\examples{
\dontrun{
quiet(cat("test"))
quiet(warning("test"))
quiet(warning("test"), all=T)
}
# This is a function that suppresses log info
}
