% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_files.R
\name{import_files}
\alias{import_files}
\title{import_files}
\usage{
import_files(path, filetype = "tsv", col_names = TRUE)
}
\arguments{
\item{path}{path where the files are present}

\item{filetype}{example: "tsv" "csv", default set to tsv}

\item{col_names}{TRUE if first row is the header, the default}
}
\description{
imports files
#'
}
\examples{
\dontrun{
quiet(cat("test"))
quiet(warning("test"))
quiet(warning("test"), all=T)
}
# This is a function that suppresses log info
}
