% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tsv.R
\name{list_tsv}
\alias{list_tsv}
\title{list_tsv}
\usage{
list_tsv(flnm, col_names2 = TRUE)
}
\arguments{
\item{flnm}{name of the file}

\item{col_names2}{TRUE if first row is the header, the default}
}
\description{
This is a function that suppresses log info
#'
}
\examples{
\dontrun{
quiet(cat("test"))
quiet(warning("test"))
quiet(warning("test"), all=T)
}
# This is a function that suppresses log info
}
